Refactor: update `RemainingWidgets` folder & files structure
