--- src/modules/dashboard/components/CoinMachine/RemainingDisplayWidget/index.ts
@@ -1,5 +1,0 @@
-export {
-  default,
-  DataDisplayType,
-  PeriodTokens as PeriodTokensType,
-} from './RemainingDisplayWidget';
