+++ src/modules/dashboard/components/CoinMachine/RemainingDisplayWidgets/RemainingDisplayWidget.css
