+++ src/modules/dashboard/components/CoinMachine/RemainingDisplayWidgets/RemainingDisplayWidget.tsx
@@ -1,0 +1,81 @@
+import React from 'react';
+import { FormattedMessage } from 'react-intl';
+import classnames from 'classnames';
+
+import Heading from '~core/Heading';
+import QuestionMarkTooltip from '~core/QuestionMarkTooltip';
+
+import { getMainClasses } from '~utils/css';
+
+import TokenPriceStatusIcon from '../TokenPriceStatusIcon/TokenPriceStatusIcon';
+
+import styles from './RemainingDisplayWidget.css';
+
+export enum DataDisplayType {
+  Time = 'Time',
+  Tokens = ' Tokens',
+}
+
+type Appearance = {
+  theme?: 'white' | 'danger';
+};
+
+type Props = {
+  appearance?: Appearance;
+  widgetText: any;
+  isWarning: any;
+  displayedValue: any;
+  priceStatus?: any;
+};
+
+const displayName = 'dashboard.CoinMachine.RemainingDisplayWidget';
+
+const RemainingDisplayWidget = ({
+  widgetText,
+  appearance = { theme: 'white' },
+  isWarning,
+  displayedValue,
+  priceStatus,
+}: Props) => {
+  return (
+    <div className={getMainClasses(appearance, styles)}>
+      <div className={styles.header}>
+        <Heading
+          text={widgetText.title}
+          appearance={{
+            size: 'small',
+            theme: appearance.theme === 'danger' ? 'invert' : 'dark',
+          }}
+        />
+        <QuestionMarkTooltip
+          className={styles.tooltipIcon}
+          tooltipText={widgetText.tooltipText}
+          invertedIcon={appearance.theme === 'danger'}
+          tooltipClassName={styles.tooltip}
+        />
+      </div>
+      <p
+        className={classnames(styles.value, {
+          [styles.valueWarning]: isWarning,
+        })}
+      >
+        {displayedValue}
+      </p>
+      {
+        // periodTokens &&
+        widgetText.footerText && (
+          <div className={styles.footer}>
+            <p className={styles.footerText}>
+              <FormattedMessage {...widgetText.footerText} />
+            </p>
+            {priceStatus && <TokenPriceStatusIcon status={priceStatus} />}
+          </div>
+        )
+      }
+    </div>
+  );
+};
+
+RemainingDisplayWidget.displayName = displayName;
+
+export default RemainingDisplayWidget;
