+++ src/modules/dashboard/components/CoinMachine/RemainingDisplayWidgets/index.ts
@@ -1,0 +1,2 @@
+export { default as RemainingTokens } from './RemainingTokens';
+export { default as RemainingTime } from './RemainingTime';
